requirements.txt
Flask==3.1.0
mysql-connector-python==9.1.0
bcrypt==4.2.1
PyJWT==2.10.1
Flask-Limiter==3.10.1
Flask-WTF==1.0.1